move key-gen test setup to setup module, remove unused deps
