+++ oprf-key-gen/Cargo.toml
@@ -11,6 +11,9 @@ license.workspace = true
 keywords = ["cryptography", "mpc", "oprf"]
 publish = false
 
+[package.metadata.cargo-machete]
+ignored = ["humantime-serde"]
+
 [dependencies]
 alloy = { workspace = true, features = [
   "contract",
@@ -33,10 +36,8 @@ config = { workspace = true }
 eyre.workspace = true
 futures = { workspace = true }
 groth16-material = { workspace = true, features = ["circom"] }
-humantime.workspace = true
 humantime-serde = { workspace = true }
 itertools = { workspace = true }
-k256 = { workspace = true, features = ["ecdsa", "serde"] }
 metrics = { workspace = true }
 nodes-common = { workspace = true, features = [
   "api",
@@ -45,13 +46,13 @@ nodes-common = { workspace = true, features = [
   "web3",
 ] }
 oprf-core = { package = "taceo-oprf-core", path = "../oprf-core", version = "0.6" }
-oprf-types = { package = "taceo-oprf-types", path = "../oprf-types", version = "0.14", features = ["service"] }
+oprf-types = { package = "taceo-oprf-types", path = "../oprf-types", version = "0.14", features = [
+  "service"
+] }
 rand.workspace = true
-reqwest = { workspace = true, features = ["json"] }
 rustls = { workspace = true }
 secrecy = { workspace = true, features = ["serde"] }
 serde = { workspace = true, features = ["derive"] }
-serde_json.workspace = true
 sqlx = { workspace = true, features = [
   "migrate",
   "postgres",
