+++ oprf-service/Cargo.toml
@@ -11,6 +11,9 @@ license.workspace = true
 keywords = ["cryptography", "mpc", "oprf"]
 publish = true
 
+[package.metadata.cargo-machete]
+ignored = ["humantime-serde"]
+
 [dependencies]
 alloy = { workspace = true, features = [
   "contract",
@@ -28,9 +31,7 @@ axum-extra = { workspace = true, features = ["typed-header"] }
 backon = { workspace = true, features = ["std", "tokio-sleep"] }
 ciborium = { workspace = true }
 eyre.workspace = true
-futures = { workspace = true }
 http = { workspace = true }
-humantime.workspace = true
 humantime-serde = { workspace = true }
 metrics = { workspace = true }
 moka.workspace = true
